Moved new utilities to proper `utils` folder
